# docs/openapi/trivida-openapi.yaml
openapi: 3.0.3
info:
  title: Trivida API
  version: 1.0.0
  description: >-
    Backend REST API for Trivida MVP (Node/Express + MongoDB). JWT access 15 min, refresh rotation 7 days.
servers:
  - url: https://api.trivida.example.com
    description: Production
  - url: http://localhost:4000
    description: Local Dev
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CronKey:
      type: apiKey
      in: header
      name: X-Cron-Key
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { nullable: true }
    Tokens:
      type: object
      required: [accessToken, accessTokenExpiresAt, refreshToken, refreshTokenExpiresAt]
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: integer, description: unix ms }
        refreshToken: { type: string }
        refreshTokenExpiresAt: { type: integer, description: unix ms }
    User:
      type: object
      required: [_id, email, locale, currency]
      properties:
        _id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        locale: { type: string, enum: [fr, en], default: fr }
        currency: { type: string, default: XAF }
        expoPushTokens:
          type: array
          items: { type: string }
...
paths:
  /auth/register:
    post:
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
                locale: { type: string, enum: [fr, en] }
                currency: { type: string }
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  tokens: { $ref: '#/components/schemas/Tokens' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/login:
    post:
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  tokens: { $ref: '#/components/schemas/Tokens' }
        '401': { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/oauth/google:
    post:
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties:
                idToken: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/User' }, tokens: { $ref: '#/components/schemas/Tokens' } } } } } }
  /auth/oauth/apple:
    post:
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties:
                idToken: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/User' }, tokens: { $ref: '#/components/schemas/Tokens' } } } } } }
  /auth/refresh:
    post:
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { tokens: { $ref: '#/components/schemas/Tokens' } } } } } }
        '401': { description: Invalid refresh, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /cron/savings/reminders:
    post:
      security:
        - CronKey: []
      tags: [Cron]
      responses:
        '202': { description: Accepted }
  /cron/budget/alerts:
    post:
      security:
        - CronKey: []
      tags: [Cron]
      responses:
        '202': { description: Accepted }
  /cron/fx/warm:
    post:
      security:
        - CronKey: []
      tags: [Cron]
      responses:
        '202': { description: Accepted }